requirements.txt
gradio==4.36.1
huggingface_hub==0.23.4
transformers==4.41.2
sentence-transformers==2.7.0
faiss-cpu
rank-bm25
torch
numpy
beautifulsoup4
requests
groq
#reset